api_antifraude/requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6 
psycopg2-binary==2.9.9 
pydantic==2.8.2 
pyjwt==2.9.0
asyncpg==0.29.0